feat: implement cogs
